Improve execution of notmuch
We streamline the execution of the notmuch cli client to handle in a common way
all reading of stdout and stderr, as well as the proper cleanup of a dead
process.

This also puts all common code in one place and lets the caller handle just the
parts that are specific to it.