Handle a message that doesn't end with a newline
A spam message appeared to have no newline at the end of its mime part, this
caused the part end markup to appear at the end of the line rather than at the
start of a line as we expected.

We now handle the case that a code marker appears at the end of the line. It
can't appear at the middle since the code includes a newline itself.